requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1


# backwards compatibility
pathlib2

# our packages
transformers
pandas
torch
pytest
hydra-core
pytorch-lightning
wandb
